+++ updater.py
@@ -55,8 +55,9 @@ try:
     )
 except Exception:
     APP_DISPLAY_NAME = "Radio & TV Segmenter"
-    PROJECT_VERSION = "2.9.1"
+    PROJECT_VERSION = "2.9.2"
     DEFAULT_GITHUB_REPO = "bradlinder/RTVS"
+
     INTERNAL_APP_ID = "RadioTVStorySegmenter"
 
     def get_github_repo() -> str: